requirements.txt
django
djangorestframework
django-webpack-loader
django-cors-headers
setuptools
twine